SSSSearch: move sstructure searches to separate file
