+++ src/glados/api/sssearch/blast.py
@@ -1,0 +1,5 @@
+def queue_blast_job():
+    response = {
+        'search_id': 'blast_id'
+    }
+    return response
