--- src/glados/api/sssearch/structure_and_sequence_search_manager.py
@@ -1,333 +1,0 @@
-# this handles the management of the special searches, structure (similarity, substructure, connectivity) and
-# sequence (blast)
-from glados.models import SSSearchJob
-import json
-import hashlib
-import base64
-from glados import glados_server_statistics
-import datetime
-import socket
-from django_rq import job
-import time
-import traceback
-import requests
-import os
-from django.conf import settings
-from glados.settings import RunEnvs
-import re
-from django.http import JsonResponse, HttpResponse
-import subprocess
-
-
-
-class SSSearchError(Exception):
-    """Base class for exceptions in this file."""
-    pass
-
-
-# ----------------------------------------------------------------------------------------------------------------------
-# Entry functions
-# ----------------------------------------------------------------------------------------------------------------------
-def request_submit_substructure_search(request):
-
-    if request.method == "POST":
-
-        search_type = request.POST.get('search_type')
-        raw_search_params = request.POST.get('raw_search_params')
-
-        try:
-
-            response = generate_search_job(search_type, raw_search_params)
-            return JsonResponse(response)
-
-        except Exception as e:
-
-            traceback.print_exc()
-            return HttpResponse("Internal Server Error: {}".format(repr(e)), status=500)
-
-    else:
-        return JsonResponse({'error': 'this is only available via POST! You crazy hacker! :P'})
-
-
-def request_sssearch_status(request, search_id):
-
-    if request.method != "GET":
-        return JsonResponse({'error': 'This is only available via GET'})
-
-    try:
-        response = get_sssearch_status(search_id)
-        return JsonResponse(response)
-    except Exception as e:
-        traceback.print_exc()
-        return HttpResponse('Internal Server Error', status=500)
-
-# ----------------------------------------------------------------------------------------------------------------------
-# results size limit
-# ----------------------------------------------------------------------------------------------------------------------
-WEB_RESULTS_SIZE_LIMIT = settings.FILTER_QUERY_MAX_CLAUSES
-CONTEXT_PREFIX = '_context'
-
-
-def get_results_file_path(job_id):
-    return os.path.join(settings.SSSEARCH_RESULTS_DIR, job_id + '.json')
-
-
-def get_search_id(search_type, raw_search_params):
-
-    stable_raw_search_params = json.dumps(json.loads(raw_search_params), sort_keys=True)
-    search_params_digest = hashlib.sha256(stable_raw_search_params.encode('utf-8')).digest()
-    base64_search_params_digest = base64.b64encode(search_params_digest).decode('utf-8').replace('/', '_').replace('+', '-')
-
-    job_id = '{}-{}'.format(search_type, base64_search_params_digest)
-    return job_id
-
-
-def get_initial_search_url(search_params, search_job):
-
-    search_term = search_params['search_term']
-    page_size = 1000
-    append_to_job_log(search_job, 'page_size: {}'.format(page_size))
-    search_type = search_job.search_type
-
-    if search_type == SSSearchJob.SIMILARITY:
-
-        threshold = search_params['threshold']
-        search_url = 'https://www.ebi.ac.uk/chembl/api/data/similarity/{search_term}/{threshold}.json' \
-                     '?limit={page_size}&only=molecule_chembl_id,similarity'.format(search_term=search_term,
-                                                                                    threshold=threshold,
-                                                                                    page_size=page_size)
-
-    elif search_type == SSSearchJob.SUBSTRUCTURE:
-
-        search_url = 'https://www.ebi.ac.uk/chembl/api/data/substructure/{search_term}.json' \
-                     '?limit={page_size}&only=molecule_chembl_id'.format(search_term=search_term, page_size=page_size)
-
-    elif search_type == SSSearchJob.CONNECTIVITY:
-
-        search_url = 'https://www.ebi.ac.uk/chembl/api/data/molecule.json?limit={page_size}' \
-                     '&only=molecule_chembl_id' \
-                     '&molecule_structures__canonical_smiles__flexmatch={search_term}'.format(
-                      search_term=search_term, page_size=page_size
-                     )
-
-    return search_url
-
-
-def append_to_results_from_response_page(response, results, search_type):
-
-    if search_type == SSSearchJob.SIMILARITY:
-
-        for r in response['molecules']:
-            results.append({
-                'molecule_chembl_id': r['molecule_chembl_id'],
-                'similarity': float(r['similarity'])
-            })
-
-    elif search_type == SSSearchJob.SUBSTRUCTURE or search_type == SSSearchJob.CONNECTIVITY:
-
-        for r in response['molecules']:
-            results.append(r)
-
-
-@job
-def do_structure_search(job_id):
-
-    start_time = time.time()
-    search_job = SSSearchJob.objects.get(search_id=job_id)
-    search_job.status = SSSearchJob.SEARCHING
-    search_job.worker = socket.gethostname()
-    search_job.save()
-    append_to_job_log(search_job, 'Performing Search')
-    print('processing job: ', job_id)
-
-    raw_search_params = search_job.raw_search_params
-    parsed_search_params = json.loads(raw_search_params)
-
-    try:
-
-        search_url = get_initial_search_url(parsed_search_params, search_job)
-        page_num = 1
-        results = []
-        # receive the results and then save them to the context file
-        # I will save all the results sorted as I receive them. It is responsibility of anyone who reads the
-        # results to load only a subset if necessary. For example, loading only the first 10000.
-        more_results_to_load = True
-        while more_results_to_load:
-
-            append_to_job_log(search_job, 'loading page: {} url: {}'.format(page_num, search_url))
-            r = requests.get(search_url)
-            response = r.json()
-            append_to_results_from_response_page(response, results, search_job.search_type)
-
-            next_page = response['page_meta']['next']
-            if next_page is not None:
-                search_url = 'https://www.ebi.ac.uk{}'.format(next_page)
-                page_num += 1
-            else:
-                more_results_to_load = False
-
-        output_file_path = get_results_file_path(job_id)
-        append_to_job_log(search_job, 'output_file_path: {}'.format(output_file_path))
-        with open(output_file_path, 'w') as outfile:
-            json.dump(results, outfile)
-
-        if settings.RUN_ENV == RunEnvs.PROD:
-            rsync_to_the_other_nfs(search_job)
-
-        append_to_job_log(search_job, 'Results Ready')
-        save_search_job_state(search_job, SSSearchJob.FINISHED)
-
-        end_time = time.time()
-        time_taken = end_time - start_time
-        glados_server_statistics.record_search(search_job.search_type, time_taken)
-
-    except:
-        save_search_job_state(search_job, SSSearchJob.ERROR)
-        tb = traceback.format_exc()
-        append_to_job_log(search_job, "Error:\n{}".format(tb))
-        print(tb)
-        return
-
-
-def generate_search_job(search_type, raw_search_params):
-
-    search_types = [s[0] for s in SSSearchJob.SEARCH_TYPES]
-    if search_type not in search_types:
-
-        raise SSSearchError(
-            "search_type: {} is unknown. Possible types are: {}".format(search_type, ', '.join(search_types))
-        )
-
-    job_id = get_search_id(search_type, raw_search_params)
-
-    try:
-        sssearch_job = SSSearchJob.objects.get(search_id=job_id)
-        if sssearch_job.status == SSSearchJob.ERROR:
-            # if it was in error state, requeue it
-            sssearch_job.status = SSSearchJob.SEARCH_QUEUED
-            sssearch_job.save()
-            append_to_job_log(sssearch_job, "Job was in error state, queuing again.")
-            do_structure_search.delay(job_id)
-
-        elif sssearch_job.status == SSSearchJob.FINISHED:
-
-            # check that the results file exists and save statistics, if not, queue it again
-            try:
-
-                file_path = get_results_file_path(job_id)
-                file_size = os.path.getsize(file_path)
-                glados_server_statistics.record_search(sssearch_job.search_type, 0, False)
-
-            except FileNotFoundError:
-
-                sssearch_job.status = SSSearchJob.SEARCH_QUEUED
-                sssearch_job.save()
-                append_to_job_log(sssearch_job, "Resuls File not found, queuing again.")
-                do_structure_search.delay(job_id)
-
-    except SSSearchJob.DoesNotExist:
-
-        sssearch_job = SSSearchJob(
-            search_id=job_id,
-            search_type=search_type,
-            raw_search_params=raw_search_params,
-            log=format_log_message('Job Queued')
-        )
-        sssearch_job.save()
-        do_structure_search.delay(job_id)
-
-    response = {
-        'search_id': job_id
-    }
-    return response
-
-
-def get_sssearch_status(search_id):
-
-    try:
-        sssearch_job = SSSearchJob.objects.get(search_id=search_id)
-        response = {
-            'status': sssearch_job.status
-        }
-        if sssearch_job.status == SSSearchJob.FINISHED:
-            context, total_results = get_search_results_context(sssearch_job)
-            response['ids'] = [k['molecule_chembl_id'] for k in context]
-            response['total_results'] = total_results
-            response['size_limit'] = WEB_RESULTS_SIZE_LIMIT
-
-        return response
-
-    except SSSearchJob.DoesNotExist:
-
-        response = {
-            'msg': 'search job does not exist!',
-            'status': SSSearchJob.ERROR
-        }
-        return response
-
-    except FileNotFoundError:
-
-        response = {
-            'msg': 'Search results not found, they may have expired. Please run the search again.',
-            'status': SSSearchJob.ERROR
-        }
-        return response
-
-
-# ----------------------------------------------------------------------------------------------------------------------
-# Loading context
-# ----------------------------------------------------------------------------------------------------------------------
-def get_search_results_context(sssearch_job, limited=True):
-
-    results_file_path = get_results_file_path(sssearch_job.search_id)
-    with open(results_file_path) as f:
-        context = json.load(f)
-
-    total_results = len(context)
-    if total_results > WEB_RESULTS_SIZE_LIMIT and limited:
-        context = context[0:WEB_RESULTS_SIZE_LIMIT]
-
-    return context, total_results
-
-
-# ----------------------------------------------------------------------------------------------------------------------
-# Syncing nfs
-# ----------------------------------------------------------------------------------------------------------------------
-def rsync_to_the_other_nfs(search_job):
-
-    hostname = socket.gethostname()
-    if bool(re.match("wp-p1m.*", hostname)):
-        rsync_destination_server = 'wp-p2m-54'
-    else:
-        rsync_destination_server = 'wp-p1m-54'
-
-    file_path = get_results_file_path(search_job.search_id)
-    rsync_destination = "{server}:{path}".format(server=rsync_destination_server, path=file_path)
-    rsync_command = "rsync -v {source} {destination}".format(source=file_path, destination=rsync_destination)
-    rsync_command_parts = rsync_command.split(' ')
-
-    append_to_job_log(search_job, "Rsyncing: {}".format(rsync_command))
-    subprocess.check_call(rsync_command_parts)
-
-
-# ----------------------------------------------------------------------------------------------------------------------
-# Logging to job and saving state
-# ----------------------------------------------------------------------------------------------------------------------
-def append_to_job_log(search_job, msg):
-
-    if search_job.log is None:
-        search_job.log = ''
-
-    search_job.log += format_log_message(msg)
-    search_job.save()
-
-
-def format_log_message(msg):
-    now = datetime.datetime.now()
-    return "[{date}] {hostname}: {msg}\n".format(date=now, hostname=socket.gethostname(), msg=msg)
-
-
-def save_search_job_state(search_job, new_state):
-    search_job.status = new_state
-    search_job.save()
-
